+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.1.0 - Queue functionality
+
+* Added a queue function to redeem currency for
+* Bugfixes
+* Cleaning up code
+
 ## 0.0.2 - Bugfixes
 
 * Fix AddToUser notification message